pyproject.toml
[project]
name = "class04-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.7.2",
    "openai>=1.106.1",
    "openai-agents>=0.2.11",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
